+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ master
 ~~~~~
 
 a) [TASK][BREAKING] Rename getCurrentInstance to getLocalInstance
+b) [FEATURE] AddConfiguration->getLocalHost method.
 
 
 2.0.0